renamed the old interactive script name and added a parameterized (ArgumentParser) version
